+++ src/accountTransactions.ts
@@ -5,7 +5,7 @@ import {
 } from './serialization';
 import {
     encodeWord64,
-    encodeMemo,
+    encodeDataBlob,
     encodeWord32,
     packBufferWithWord32Length,
     packBufferWithWord16Length,
@@ -25,6 +25,7 @@ import {
     UpdateContractPayload,
     AccountTransactionPayload,
     UpdateCredentialsPayload,
+    RegisterDataPayload,
 } from './types';
 
 interface AccountTransactionHandler<
@@ -54,7 +55,7 @@ export class SimpleTransferWithMemoHandler
 {
     serialize(transfer: SimpleTransferWithMemoPayload): Buffer {
         const serializedToAddress = transfer.toAddress.decodedAddress;
-        const serializedMemo = encodeMemo(transfer.memo);
+        const serializedMemo = encodeDataBlob(transfer.memo);
         const serializedAmount = encodeWord64(transfer.amount.microGtuAmount);
         return Buffer.concat([
             serializedToAddress,
@@ -72,9 +73,9 @@ export class DeployModuleHandler
         return BigInt(cost);
     }
 
-    serialize(transfer: DeployModulePayload): Buffer {
-        const serializedWasm = packBufferWithWord32Length(transfer.content);
-        const serializedVersion = encodeWord32(transfer.version);
+    serialize(payload: DeployModulePayload): Buffer {
+        const serializedWasm = packBufferWithWord32Length(payload.content);
+        const serializedVersion = encodeWord32(payload.version);
         return Buffer.concat([serializedVersion, serializedWasm]);
     }
 }
@@ -176,7 +177,7 @@ export class TransferWithScheduleAndMemoHandler
     implements AccountTransactionHandler<TransferWithScheduleAndMemoPayload>
 {
     serialize(scheduledTransfer: TransferWithScheduleAndMemoPayload): Buffer {
-        const serializedMemo = encodeMemo(scheduledTransfer.memo);
+        const serializedMemo = encodeDataBlob(scheduledTransfer.memo);
         const serializedToAddress = scheduledTransfer.toAddress.decodedAddress;
         const serializedScheduleLength = encodeWord8(
             scheduledTransfer.schedule.length
@@ -219,7 +220,7 @@ export class EncryptedTransferWithMemoHandler
 {
     serialize(encryptedTransfer: EncryptedTransferWithMemoPayload): Buffer {
         const serializedToAddress = encryptedTransfer.toAddress.decodedAddress;
-        const serializedMemo = encodeMemo(encryptedTransfer.memo);
+        const serializedMemo = encodeDataBlob(encryptedTransfer.memo);
         const serializedEncryptedData =
             serializeEncryptedData(encryptedTransfer);
 
@@ -275,6 +276,18 @@ export class UpdateCredentialsHandler
     }
 }
 
+export class RegisterDataHandler
+    implements AccountTransactionHandler<RegisterDataPayload>
+{
+    getBaseEnergyCost(): bigint {
+        return 300n;
+    }
+
+    serialize(payload: RegisterDataPayload): Buffer {
+        return encodeDataBlob(payload.data);
+    }
+}
+
 export function getAccountTransactionHandler(
     type: AccountTransactionType.SimpleTransfer
 ): SimpleTransferHandler;
@@ -294,9 +307,6 @@ export function getAccountTransactionHandler(
     type: AccountTransactionType.EncryptedTransferWithMemo
 ): EncryptedTransferWithMemoHandler;
 export function getAccountTransactionHandler(
-    type: AccountTransactionType
-): AccountTransactionHandler;
-export function getAccountTransactionHandler(
     type: AccountTransactionType.UpdateCredentials
 ): UpdateCredentialsHandler;
 export function getAccountTransactionHandler(
@@ -311,6 +321,13 @@ export function getAccountTransactionHandler(
 export function getAccountTransactionHandler(
     type: AccountTransactionType.UpdateSmartContractInstance
 ): UpdateContractHandler;
+export function getAccountTransactionHandler(
+    type: AccountTransactionType.RegisterData
+): RegisterDataHandler;
+
+export function getAccountTransactionHandler(
+    type: AccountTransactionType
+): AccountTransactionHandler;
 
 // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
 export function getAccountTransactionHandler(type: AccountTransactionType) {
@@ -335,6 +352,8 @@ export function getAccountTransactionHandler(type: AccountTransactionType) {
             return new UpdateContractHandler();
         case AccountTransactionType.UpdateCredentials:
             return new UpdateCredentialsHandler();
+        case AccountTransactionType.RegisterData:
+            return new RegisterDataHandler();
         default:
             throw new Error(
                 'The provided type does not have a handler: ' + type